# openapi/event-gateway.yaml
openapi: 3.0.1
info:
  title: Event Gateway API
  description: "Public-facing entry point for submitting transaction events. Enforces\
    \ idempotency on eventId, tolerates out-of-order arrival, stores every event locally\
    \ as an audit record, and degrades gracefully (never hangs, never 500s) when the\
    \ internal Account Service is unreachable."
  contact:
    name: Event Ledger Team
  version: v1
servers:
- url: http://localhost:8082
  description: Local
tags:
- name: Events
  description: Submit and query transaction events (public entry point)
- name: Accounts (proxy)
  description: Pass-through balance lookups to the internal Account Service
paths:
  /events:
    get:
      tags:
      - Events
      summary: "List events for an account, chronologically ordered"
      description: "Ordered by the event's own business timestamp, not arrival order\
        \ — a late-arriving event with an earlier timestamp is placed correctly. Served\
        \ entirely from the Gateway's own database."
      operationId: listByAccount
      parameters:
      - name: account
        in: query
        description: Account to list events for
        required: true
        schema:
          type: string
        example: acct-123
      responses:
        "200":
          description: Events returned (possibly empty)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
    post:
      tags:
      - Events
      summary: Submit a transaction event
      description: "Idempotent on eventId. Tolerates out-of-order arrival (balance\
        \ is always order-independent). If the Account Service is unreachable, the\
        \ event is still recorded locally (status FAILED) and 503 is returned — never\
        \ a hang, never a 500."
      operationId: submit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
        required: true
      responses:
        "201":
          description: "New event, successfully applied"
          content:
            '*/*':
              schema:
                type: object
        "503":
          description: Account Service unreachable; event recorded with status FAILED
            for later retry
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: "Validation failed (missing field, non-positive amount, unknown\
            \ type)"
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Duplicate of an already-applied eventId — original record returned
            unchanged
          content:
            '*/*':
              schema:
                type: object
  /events/{id}:
    get:
      tags:
      - Events
      summary: Retrieve a single event by its ID
      description: "Served entirely from the Gateway's own database — never depends\
        \ on the Account Service, so it keeps working even during an outage."
      operationId: getById
      parameters:
      - name: id
        in: path
        description: The eventId supplied at submission time
        required: true
        schema:
          type: string
        example: evt-001
      responses:
        "404":
          description: No event with this eventId has been submitted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponse'
        "200":
          description: Event found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponse'
  /accounts/{accountId}/balance:
    get:
      tags:
      - Accounts (proxy)
      summary: Get the current balance for an account
      description: Proxies to the Account Service through the same circuit-breaker/timeout/retry
        wrapping used for event submission. Returns 503 with a clear message if the
        Account Service is unreachable — never a fabricated or stale balance.
      operationId: getBalance
      parameters:
      - name: accountId
        in: path
        description: Account to look up
        required: true
        schema:
          type: string
        example: acct-123
      responses:
        "404":
          description: Account has never received a transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        "200":
          description: Balance returned
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        "503":
          description: Account Service unreachable
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceResponse'
components:
  schemas:
    EventRequest:
      required:
      - accountId
      - amount
      - currency
      - eventId
      - eventTimestamp
      - type
      type: object
      properties:
        eventId:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum:
          - CREDIT
          - DEBIT
        amount:
          minimum: 0.0
          exclusiveMinimum: true
          type: number
        currency:
          type: string
        eventTimestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: object
    EventResponse:
      type: object
      properties:
        eventId:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum:
          - CREDIT
          - DEBIT
        amount:
          type: number
        currency:
          type: string
        eventTimestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: object
        status:
          type: string
          enum:
          - RECEIVED
          - APPLIED
          - FAILED
        failureReason:
          type: string
        createdAt:
          type: string
          format: date-time
    BalanceResponse:
      type: object
      properties:
        accountId:
          type: string
        balance:
          type: number